Change: Implements http codes response, and move folder logs

+++ src/main/java/com/retotecnico/ApiBackend/controller/PersonaController.java
@@ -1,5 +1,7 @@
 package com.retotecnico.ApiBackend.controller;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.retotecnico.ApiBackend.entity.Persona;
 import com.retotecnico.ApiBackend.model.ErrorModel;
 import com.retotecnico.ApiBackend.service.PersonaService;
@@ -17,44 +19,78 @@ import java.util.List;
 @RestController
 @RequestMapping("/api-persona")
 public class PersonaController {
+
+    private final ObjectMapper objectMapper;
     @Autowired
     protected PersonaService personaService;
 
+    public PersonaController(ObjectMapper objectMapper) {
+        this.objectMapper = objectMapper;
+    }
+
     @PostMapping("/save")
-    public Persona save(@Valid @RequestBody Persona persona){
-        log.error("TEST de error");
-        log.info("Test de info");
-        return personaService.save(persona);
+    public ResponseEntity<?> save(@Valid @RequestBody Persona persona){
+        try{
+            log.info("Request Body: {}",objectMapper.writeValueAsString(persona));
+            ResponseEntity<?> response =  ResponseEntity.ok(personaService.save(persona));
+            log.info(response.getStatusCode() + " Response Body: {}",objectMapper.writeValueAsString(response.getBody()));
+            return response;
+        }catch (Exception e){
+            log.error("Error guardando la persona: {}" + e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error guardando la persona");
+        }
     }
 
     @GetMapping("/list-all")
-    public List<Persona> findAll(){
-        return personaService.findAll();
+    public ResponseEntity<?> findAll(){
+        List<Persona> list = personaService.findAll();
+        if(!list.isEmpty()) return ResponseEntity.ok(personaService.findAll());
+        else return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No se encuentran personas registradas");
     }
 
     @GetMapping("/list")
-    public Persona findById(@RequestParam("id") Integer id){
-        return personaService.findById(id);
+    public ResponseEntity<?> findById(@RequestParam("id") Integer id){
+        Persona persona = personaService.findById(id);
+        if (persona != null)
+            return ResponseEntity.ok(persona);
+        else return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Persona no registrada");
     }
     @GetMapping("/list/{id}")
-    public Persona findByIdPath(@PathVariable("id") Integer id){
-        return personaService.findById(id);
+    public ResponseEntity<?> findByIdPath(@PathVariable("id") Integer id){
+        Persona persona = personaService.findById(id);
+        if (persona != null)
+            return ResponseEntity.ok(persona);
+        else return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Persona no registrada");
     }
     @PutMapping("/edit/{id}")
     public ResponseEntity<?> editPersona(@RequestBody Persona persona, @PathVariable("id") Integer id){
-        if (findById(id) != null){
+        Persona personaFind = personaService.findById(id);
+        if (personaFind != null){
             persona.setId(id);
-            return new ResponseEntity<>(personaService.editPersona(persona), HttpStatus.OK);
+            try{
+                log.info("Request Body: {}",objectMapper.writeValueAsString(persona));
+                ResponseEntity<?> response =  ResponseEntity.ok(personaService.editPersona(persona));
+                log.info(response.getStatusCode() + " Response Body: {}",objectMapper.writeValueAsString(response.getBody()));
+                return response;
+            }catch (Exception e){
+                log.error("Error editando la persona: {}" + e);
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error editando la persona");
+            }
         }
-        else return new ResponseEntity<>("Elemento no encontrado",HttpStatus.NOT_FOUND);
+        else return new ResponseEntity<>("Persona no registrada",HttpStatus.NOT_FOUND);
     }
     @GetMapping("/truncate")
     public void truncate(){
         personaService.truncateTable();
     }
     @DeleteMapping("/delete/{id}")
-    public void deleteById(@PathVariable("id") Integer id){
-        personaService.deletePersona(id);
+    public ResponseEntity<?> deleteById(@PathVariable("id") Integer id){
+        Persona persona = personaService.findById(id);
+        if (persona != null){
+            personaService.deletePersona(id);
+            return ResponseEntity.ok("Registro eliminado: "+persona.getNombres());
+        }
+        else return new ResponseEntity<>("Persona no registrada",HttpStatus.NOT_FOUND);
     }
 
     @ExceptionHandler(MethodArgumentNotValidException.class)
@@ -65,6 +101,11 @@ public class PersonaController {
         errorModel.setErrorType("Campos Obligatorios");
         assert fieldError != null;
         errorModel.setErrorDescription(fieldError.getDefaultMessage());
+        try {
+            log.error(" Response Body: {}", objectMapper.writeValueAsString(errorModel));
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorModel);
     }
 }
